+++ node/clients/catalog.ts
@@ -3,7 +3,12 @@ import { JanusClient } from '@vtex/api'
 
 export default class Catalog extends JanusClient {
   constructor(ctx: IOContext, options?: InstanceOptions) {
-    super(ctx, { ...options })
+    super(ctx, {
+      ...options,
+      headers: {
+        VtexIdClientAutCookie: ctx.authToken,
+      },
+    })
   }
 
   public async salesChannelAvailable(email: string) {
@@ -13,6 +18,15 @@ export default class Catalog extends JanusClient {
       )}`,
       {
         metric: 'catalog-sales-channel-available',
+      }
+    )
+  }
+
+  public async collectionsAvailable(searchTerm: string) {
+    return this.http.get<PagedCollectionsResponseAPI>(
+      `/api/catalog_system/pvt/collection/search/${encodeURI(searchTerm)}`,
+      {
+        metric: 'catalog-collections-available',
       }
     )
   }
@@ -33,4 +47,28 @@ interface SalesChannelAvailable {
   Position: number
   ConditionRule: string | null
   CurrencyDecimalDigits: null | number
+}
+
+interface Paging {
+  pages: number
+  page: number
+  perPage: number
+  total: number
+}
+
+interface CollectionItem {
+  id: number
+  name: string
+  searchable: boolean
+  highlight: boolean
+  dateFrom: string
+  dateTo: string
+  totalProducts: number
+  type: 'Hybrid' | 'Manual' | 'Automatic'
+  lastModifiedBy: string | null
+}
+
+interface PagedCollectionsResponseAPI {
+  paging: Paging
+  items: CollectionItem[]
 }